Final check forced unwrapping & Error Handling, Remove 불필요한 주석 및 print 문 & 앨범이 6개 일 시 추가 공유 안되도록
